refactor: extract test modules and add test infrastructure
- Extract tests from 5 major source files to separate *_tests.rs files:
  emitter (1724 lines), parser (585), sync (1383), tools (1005), layout (689)
- Add roundtrip! macro in test_helpers.rs for DRY roundtrip tests
- Add proptest roundtrip (generate random FD → parse → emit → reparse)
- Add justfile with smoke/test/extended targets for tiered execution
- Add .config/nextest.toml for parallel test runner
- Source files reduced ~4,800 lines (production code only)
- All 303 tests pass (302 unit + 1 proptest)
