+++ crates/fd-core/Cargo.toml
@@ -21,3 +21,4 @@ log = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = { workspace = true }
+proptest = "1"
